+++ iOS/Hummer/Classes/Component/Views/TextArea/HMTextArea.m
@@ -15,10 +15,11 @@
 #import "HMInputEvent.h"
 #import "UIView+HMEvent.h"
 #import "UIView+HMDom.h"
+#import <Hummer/UIView+HMInspector.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface HMTextArea() <UITextViewDelegate>
+@interface HMTextArea() <UITextViewDelegate, HMViewInspectorDescription>
 
 @property (nonatomic, strong, nullable) UILabel *placeholderLabel;
 
@@ -285,6 +286,18 @@ HM_EXPORT_ATTRIBUTE(textLineClamp, numberOfLines, HMNumberToNSInteger:)
         }
     }
     [self togglePlaceholder];
+}
+
+
+#pragma mark - <HMDescription>
+- (NSString *)hm_content {
+    
+    return self.text;
+}
+
+- (nullable NSArray<id<HMViewInspectorDescription>> *)hm_displayJsChildren {
+    
+    return nil;
 }
 
 @end