+++ iOS/Hummer/Classes/Component/WebSocket/HMWebSocket.m
@@ -74,7 +74,7 @@ HM_EXPORT_METHOD(send, send:)
     }
     // 如果当前连接尚未准备好，则忽略错误，或者抛出 JS 异常
     // 正常应当从 onopen 调用后开始 send
-    [self.webSocket sendString:dataString error:nil];
+    [self.webSocket send:dataString];
 }
 
 - (instancetype)initWithHMValues:(NSArray<__kindof HMBaseValue *> *)values {